Refactor to make this seasons table file not gigantic
