+++ src/components/tables/SeasonsTable.tsx
@@ -5,11 +5,13 @@ import { SeasonsTableData } from "@/state/useSeasonsData";
 import { Tooltip, TooltipProvider, TooltipTrigger } from "@/components/ui/Tooltip";
 import { TooltipContent, TooltipPortal } from "@radix-ui/react-tooltip";
 import { seasonColumns, SortColumn } from "@/pages/explorer/SeasonsExplorer";
-import { Separator } from "@/components/ui/Separator";
 import { TokenValue } from "@/classes/TokenValue";
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { areEqual, ListChildComponentProps, VariableSizeList } from "react-window";
 import useIsMobile from "@/hooks/display/useIsMobile";
+import { calculateCropScales } from "@/utils/convert";
+import { caseIdToDescriptiveText } from "@/utils/utils";
+
 enum SeasonsTableCellType {
   Default = "default",
   TwoColumn = "twoColumn",
@@ -24,48 +26,6 @@ interface SeasonsTableCellProps {
   hoverContent?: any;
 }
 
-// biome-ignore lint/correctness/noUnusedVariables: <explanation>
-const DeltaDemandChart = () => {
-  const lowerBound = 15000;
-  const upperBound = 15000;
-  const secondLowerBound = 24000;
-  const secondUpperBound = 24000;
-  return (
-    <div className="w-[600px] bg-white rounded-md p-6 border border-pinto-gray-2">
-      <span>Demand for Soil is increasing</span>
-      <div className="flex text-sm mt-2 gap-1 items-center">
-        <div className="rounded-full w-3 h-3 bg-pinto-green-4" />
-        <span>Season 998</span>
-        <span className="text-pinto-gray-4">(33% Temp)</span>
-        <span>-</span>
-        <span className="text-pinto-gray-4">Amount Sown:</span>
-        <span>
-          {lowerBound}/{upperBound}
-        </span>
-        <span className="text-pinto-gray-4">(100%)</span>
-      </div>
-      <div className="flex text-sm mt-2 gap-1 items-center">
-        <div className="rounded-full w-3 h-3 bg-pinto-morning-yellow-1" />
-        <span>Season 999</span>
-        <span className="text-pinto-gray-4">(32% Temp)</span>
-        <span>-</span>
-        <span className="text-pinto-gray-4">Amount Sown:</span>
-        <span>
-          {secondLowerBound}/{secondUpperBound}
-        </span>
-        <span className="text-pinto-gray-4">(100%)</span>
-      </div>
-      <div className="w-full h-[200px] bg-pinto-gray-1 rounded-md mt-2">Super Slick Chart here</div>
-      <Separator className="my-4" />
-      <div className="flex flex-col">
-        <span className="text-base">Demand for Soil</span>
-        <span className="text-xl">Increasing</span>
-        <span className="text-base mt-4 text-pinto-gray-4">Soil - 1 Sown a Season 998: XX:20</span>
-        <span className="text-base text-pinto-gray-4">Soil - 1 Sown a Season 997: XX:25</span>
-      </div>
-    </div>
-  );
-};
 const TwoColumnCell = ({
   className,
   value,
@@ -96,27 +56,6 @@ const TwoColumnCell = ({
   );
 };
 
-const calculateCropScales = (value: number, isRaining: boolean) => {
-  const minInput = 0;
-  const maxInput = 1e18 * 100; // value is 1-100 not 0.0 -  1.0
-  const maxOutput = 100;
-
-  // Calculate crop scalar
-  const scalarMinOutput = 0;
-  const scalarValue = scalarMinOutput + ((value - minInput) * (maxOutput - scalarMinOutput)) / (maxInput - minInput);
-  const cropScalar = (Math.min(Math.max(scalarValue, scalarMinOutput), maxOutput)).toFixed(1);
-
-  // Calculate crop ratio
-  const ratioMinOutput = isRaining ? 33 : 50;
-  const ratioValue = ratioMinOutput + ((value - minInput) * (maxOutput - ratioMinOutput)) / (maxInput - minInput);
-  const cropRatio = Math.min(Math.max(ratioValue, ratioMinOutput), maxOutput).toFixed(1);
-
-  return {
-    cropScalar,
-    cropRatio,
-  };
-};
-
 const convertDeltaDemandToPercentage = (deltaDemand: number) => {
   if (deltaDemand === 0) return "0%";
   if (deltaDemand === 1e18) return "100%";
@@ -126,35 +65,6 @@ const convertDeltaDemandToPercentage = (deltaDemand: number) => {
   const scaledValue = (deltaDemand / 1e18) * 100;
   return `${TokenValue.fromHuman(scaledValue, 0).toHuman("short")}%`;
 }
-
-const caseIdToDescriptiveText = (caseId: number, column: "price" | "soil_demand" | "pod_rate" | "l2sr") => {
-  switch (column) {
-    case "price":
-      if ((caseId % 36) % 9 < 3) return "P < $1.00";
-      else if ((caseId % 36) % 9 < 6) return "P > $1.00";
-      //(caseId % 36 < 9)
-      else return "P > Q";
-    case "soil_demand":
-      if (caseId % 3 === 0) return "Decreasing";
-      else if (caseId % 3 === 1) return "Steady";
-      else return "Increasing";
-    case "pod_rate":
-      if ((caseId % 36) / 9 === 0) return "Excessively Low";
-      else if ((caseId % 36) / 9 === 1) return "Reasonably Low";
-      else if ((caseId % 36) / 9 === 2) return "Reasonably High";
-      else return "Excessively High";
-    case "l2sr":
-      if (caseId / 36 === 0) {
-        return "Extremely Low";
-      } else if (caseId / 36 === 1) {
-        return "Reasonably Low";
-      } else if (caseId / 36 === 2) {
-        return "Reasonably High";
-      } else {
-        return "Extremely High";
-      }
-  }
-};
 
 const nonHideableFields = ["season"];
